feat: CSS global y reorganizacion
