+++ src/App.jsx
@@ -1,5 +1,5 @@
 import logo from './logo.svg';
-import './App.css';
+import './styles/App.css';
 
 const App = () => (
   <div className="App">
